+++ CHANGELOG.rst
@@ -9,7 +9,9 @@ Change Log
 5.3.0
 =====
 * Added update of a gitinfo.json file in GitHub Actions (.github/workflows/main-publish.yml).
-* Fix to the API /aws/ecs/tasks_for_running/deploy to NOT match the "initial" deploy task.
+* Fix to the (API for the) reindex page to NOT match the "initial" deploy task.
+* Fix to the (API for the) redeploy and reindex pages for to straighten out blue/green
+  and staging/data for clusters/services vs task definitions.
 
 
 5.2.0